koursa/requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==6.0
djangorestframework==3.16.1
drf-yasg==1.21.11
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
simplejwt==2.0.1
sqlparse==0.5.5
typing==3.7.4.3
uritemplate==4.2.0
whitenoise==6.11.0
